+++ assets/shaders/element_v.glsl
@@ -1,10 +1,10 @@
-#version 420
+#version 320 es
 
 in vec2 a_position;
 in vec2 a_texCoord0;
 
 uniform mat4 u_projTrans;
-uniform int u_worldWidth;
+uniform float u_worldWidth;
 
 out vec2 v_texCoords;
