+++ assets/shaders/river_v.glsl
@@ -1,11 +1,11 @@
-#version 420
+#version 320 es
 
 in vec2 a_position;
 in vec2 a_texCoord0;
 in float a_width;
 
 uniform mat4 u_projTrans;
-uniform int u_worldWidth;
+uniform float u_worldWidth;
 
 out vec2 v_texCoords;
 out vec2 v_worldCoords;
